+++ sdk/resourcemanager/videoanalyzer/armvideoanalyzer/ze_generated_example_accesspolicies_client_test.go
@@ -18,7 +18,7 @@ import (
 )
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/videoanalyzer/resource-manager/Microsoft.Media/preview/2021-11-01-preview/examples/access-policy-list.json
-func ExampleAccessPoliciesClient_List() {
+func ExampleAccessPoliciesClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -30,7 +30,7 @@ func ExampleAccessPoliciesClient_List() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.List("<resource-group-name>",
+	pager := client.NewListPager("<resource-group-name>",
 		"<account-name>",
 		&armvideoanalyzer.AccessPoliciesClientListOptions{Top: to.Ptr[int32](2)})
 	for pager.More() {