+++ sdk/resourcemanager/vmwarecloudsimple/armvmwarecloudsimple/zz_generated_operations_client.go
@@ -123,10 +123,10 @@ func (client *OperationsClient) getHandleResponse(resp *http.Response) (Operatio
 	return result, nil
 }
 
-// List - Return list of operations
+// NewListPager - Return list of operations
 // If the operation fails it returns an *azcore.ResponseError type.
 // options - OperationsClientListOptions contains the optional parameters for the OperationsClient.List method.
-func (client *OperationsClient) List(options *OperationsClientListOptions) *runtime.Pager[OperationsClientListResponse] {
+func (client *OperationsClient) NewListPager(options *OperationsClientListOptions) *runtime.Pager[OperationsClientListResponse] {
 	return runtime.NewPager(runtime.PageProcessor[OperationsClientListResponse]{
 		More: func(page OperationsClientListResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0