+++ sdk/resourcemanager/webpubsub/armwebpubsub/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armwebpubsub
 
 const (
 	moduleName    = "armwebpubsub"
-	moduleVersion = "v0.3.0"
+	moduleVersion = "v0.4.0"
 )
 
 // ACLAction - Azure Networking ACL Action.