+++ sdk/resourcemanager/workloadmonitor/armworkloadmonitor/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armworkloadmonitor
 
 const (
 	moduleName    = "armworkloadmonitor"
-	moduleVersion = "v0.3.0"
+	moduleVersion = "v0.4.0"
 )
 
 // HealthState - One of health states - healthy, critical, warning, unknown, none, disabled.