+++ sdk/resourcemanager/windowsiot/armwindowsiot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 0.4.0 (2022-04-18)
+### Breaking Changes
+
+- Function `*OperationsClient.List` has been removed
+- Function `*ServicesClient.List` has been removed
+- Function `*ServicesClient.ListByResourceGroup` has been removed
+
+### Features Added
+
+- New function `*ServicesClient.NewListByResourceGroupPager(string, *ServicesClientListByResourceGroupOptions) *runtime.Pager[ServicesClientListByResourceGroupResponse]`
+- New function `*OperationsClient.NewListPager(*OperationsClientListOptions) *runtime.Pager[OperationsClientListResponse]`
+- New function `*ServicesClient.NewListPager(*ServicesClientListOptions) *runtime.Pager[ServicesClientListResponse]`
+
+
 ## 0.3.0 (2022-04-13)
 ### Breaking Changes